ramips: rt305x: Disable image generation for small devices
It was reported in FS#2110 that there are currently no images being
generated for rt305x as a result of a build breakage due to image
generation.

I've found out, that plenty of the devices have fallen of a cliff
already due to the small flash, so this patch disables image generation
for those devices.

Signed-off-by: Petr Štetiar <ynezz@true.cz>