ath79: ag71xx: make of_get_mac_address probe defer aware
of_get_mac_address now uses NVMEM under the hood, so it could now return
EPROBE_DEFER and it needs to be handled, which means, that it needs to
be propagated back to the device driver subsystem in order to be handled
as designed.

While at it, remove superfluous call of is_valid_ether_addr, as the MAC
address returned by of_get_mac_address was already checked by
is_valid_ether_addr internally.

Signed-off-by: Petr Štetiar <ynezz@true.cz>