uboot-mediatek: Add GPT table write command for Arcadyan Mozart
Add an additional bootmenu option and commands to overwrite the GPT
table with the custom one for upstream U-Boot.

This is needed as the pre-installed GPT table differ and updating the
GPT table is a requirement to install the upstream U-Boot (or the old
U-Boot will be loaded)

To make the migration to new U-Boot bootloader an user can now load the
system with mtk_uartboot and select "Load GPT Table ..." in conjunction
with other update command.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [erase group size fix]